requirements.txt
streamlit
rdkit-pypi
torch
numpy
matplotlib
